+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Disable diagnostics on the `Container` itself when `diag_owner` is set to avoid duplicate port binding
 
 ## [1.53.0] - 2026-04-15